refactor: 링크 썸네일 컴포넌트 디렉토리 이동
