+++ src/components/add/LinkThumbnail.tsx
@@ -3,16 +3,17 @@ import { css, Theme } from '@emotion/react';
 import useOpenGraphImage from '~/hooks/common/useOpenGraphImage';
 import { textEllipsisCss } from '~/styles/utils';
 
-import { IconButton } from './common/Button';
+import { IconButton } from '../common/Button';
 
-export interface LinkThumbnailMetaData {
+// TODO: /remotes/inspiration OpenGraphResponse로 대체 가능해보임
+interface LinkThumbnailMetaData {
   image?: string;
   title: string;
   url: string;
   alt?: string;
 }
 
-export interface LinkThumbnailProps {
+interface LinkThumbnailProps {
   edit?: boolean;
   thumbnail: LinkThumbnailMetaData;
   onDelete?: VoidFunction;
@@ -78,6 +79,7 @@ const linkThumbnailBoxCss = () => css`
   height: 100px;
   border-radius: 4px;
 `;
+
 const linkThumbnailContentCss = (hasImage: boolean) => (theme: Theme) =>
   css`
     padding: 16px;
